final/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
bcrypt==3.2.0
python-dotenv==0.19.0
opencv-python==4.5.3.56
dlib==19.22.0
numpy==1.21.2
scipy==1.7.1
pygame==2.0.1
requests==2.26.0
